+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - CollectionView: added default implementation for "getHTML()" method returning simple "div" element
 - CollectionView: allowed not to specify EmptyView when it's not used
 - CollectionView: allowed to render Views inside of specified child Element
+- CollectionView: EmptyView is automatically removed on "addChildView()" calls
 - Documentation: improve example usage of "OsfObservable.retrigger()" method
 - Documentation: added "async" prefix to examples of View method usages
 - Documentation: fixed formatting in various code examples